Move logic to a trait and implement that into the post builder.
